requirements.txt
python-dotenv
mysql-connector-python
alpaca-trade-api
alpaca-py